Organize runtime code under src and normalize services
