+++ services/api/tests/test_contracts.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from api.main import app
+from services.api.main import app
 
 
 def test_health_ok():
