+++ services/example_service/entrypoint.sh
