+++ tests/test_start_api_helpers.py
@@ -1,7 +1,7 @@
 import asyncio
 import types
 
-from start_api import _warm_cache, _ensure_health_endpoint
+from src.start_api import _warm_cache, _ensure_health_endpoint
 
 
 class DummyWarmer:
